requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.13
django-dump-die==0.1.7
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
importlib-metadata==7.1.0
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
msedge-selenium-tools==3.141.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
selenium==3.141.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zipp==3.18.1
gunicorn
